A change to the solver set_options() so it can accept a dictionary-like data structure as well as a string. The 'solver_str' data is no longer maintained in a solver, but a solver string can be generated with the _get_solver_string() method.
Reworking the logic for lauching solvers in pyomo.scripting.
A solver should not be constructed explicitly here unless
a serial solver is being created.  (Even then, we may want
to defer solver construction, but we'll see about that.)
This change allows me to launch a cplex solver on NEOS without
having a local cplex software license.

NOTE:  This reverts a recent change.  We MUST be able to pass in
an 'option' argument to initialize a solver.  This is required
by solver managers, who pass through solver options.  These are not
stored persistently, because solver managers allow for the specification
of solvers with string names, not just solver objects.

Misc changes in pyomo.neos to capture these and other recent changes.


git-svn-id: https://software.sandia.gov/svn/pyomo/pyomo/trunk@10517 570ccb8d-5833-0410-9ce8-cbedd0da42eb