Adding an optimization for indexed Var construction to recover performance in a common setting, which was lost following r11067:
In cases such as Var(model.s), where model.s is a large
index set that is expensive to iterate over (note that Pyomo
Set iteration is often times slower than iteration of
built-in containers), we lost some performance compared to
code in the Pyomo-4.2 release. The issue was that following
r11067, we were iterating over the index set one additional
time in order to set the domain for each index, whereas with
Pyomo-4.2, unless domain was a rule, this iteration could be
avoided because there was only one domain object to set on
the Var container.

This commit recovers the performance of Pyomo-4.2 for these
cases by setting the domain attribute during the initial
iteration that takes place when the _VarData objects are
being created.


git-svn-id: https://software.sandia.gov/svn/pyomo/pyomo/trunk@11181 570ccb8d-5833-0410-9ce8-cbedd0da42eb