Allow mutable `Param.__getitem__` even with no default value.
This supports Concrete modelers who will create Params and then set
their values later before solving the model.  This change defers the
exception for not defining the default value until the model is written.
The side effect is that the error the user will get changes from the
Param not having a default value to a more generic param is not
initialized.