Removing the getname() overload from ExternalFunction components
The correct fix for #49 is to remove the local definition of getname()
from ExternalFunction completely.  The original implementation was a
non-standard behavior: the component would return the function
information as its name when it was not attached to a Block.  The
original fix for #49 changed it so that it would never return this
information - which is probably correct.  Having done that, the special
code for returning the function information was now unreachable and the
entire thing should just be removed.