Introducing a new factory class called 'noclone' that can be used to wrap expressions and prevent cloning. It is not a fully fledged component, so it simply uses 1 additional __slots__ to store its subexpression. It is immutable in that it can not be changed to reference a different expression (through any public methods). It requires a single argument when initialized and if that argument is not an instance of NumericValue, the argument itself is returned.
This object should be used in advanced situations where
expression cloning is causing performance issues but the
programmer knows that shared subexpressions will not created
a problem. E.g.,

def compute(x, y, z):
    e = noclone(x*y + z**2)
    return e + e**2 + e**3

The semantics of noclone in the above case are such that the
function can be called with primitive types or with
NumericValue objects and the return type will be what is
expected (that is, a primitive type in the first case and a
Pyomo expression in the second case).