Fixing pickling of ModelSolutions objects.
Overhaul of the ModelSolutions __getstate__ / __setstate__
routines. Delegating the `id()` management and updating to teh
ModelSolution object (where the ids are stored).  This also removes the
use of ComponentUIDs in the pickled stream (in favor of direct object
references that are restored to weakrefs upon unpickling).

This fixes #65 (both the exception and the performance problems).