Change the behavior of IndexedComponent_slicer.__call__
Calling an `IndexedComponent_slicer` will now immediately iterate over the
slicer and issue `__call__` on the appropriate attribute.  The only
exception will be if the attribute was `component` in which case the
previous behavior of extending the slicer (and not iterating) will
continue to occur (for backwards compatibility).  This addresses many of
the concerns raised in #155.