Rework of mutable sums
Deprecated the use of multisum types, which have been
entirely replaced with viewsum types.

Renamed _MutableViewSumExpression to _ViewSumExpression, to
reflect the fact that once created the view is static.  The
number of arguments doesn't change.

Added _MutableViewSumExpression, which is used in the
nonlinear_expression context manager to collect terms in a
summation.  While collecting these terms, the view is mutable and
we are guaranteed to not have variables pointing to these objects.
Hence, we can avoid constructing many _ViewSumExpression objects.