Several important changes
1. Fixed bugs in visitor search logic.  Now
using the expression.args property to iterate
over the just the arguments that are valid
(this impacts how ViewSum expressions are
managed).

2. Renamed clone_info to _is_owned

3. Added the _ExpressionBase.args property
and arg() method.

4. Removed cloning check in _process_arg,
and added checks for is_expression() before
calling _process_arg everywhere.

5. Remove the use of _SumClass_ everywhere.

6. If a ViewSum expression is owned, then
down share its list with the new ViewSum.
This avoids entanglement side effects.

NOTE:  Between 4 and 6, I've eliminated all dynamic
cloning in Pyomo5 expressions.  In theory, this could
lead to deeply nested sums.  But not in cases that users
are likely to use (e.g. sums generated with simple for-loops).
But those same cases would have generated excessive cloning
anyway, so this seems like a better alternative.