Reworking the Sum() function
This now peeks at the first term, checks if it's linear or not
and then assumes that the rest of the terms are the same.

This is less general, but it's captures the most common case
when using generators in models.  With this change, users
don't really need to use the linear_expression context manager
except when pulling together inconsistent expressions.