Use visitor pattern
This uses the pyomo.core.expr visitors to convert
from Pyomo to Sympy expressions, and it uses the pyutilib.misc.visitor
visitors to convert from Sympy to Pyomo.  (The latter visitor pattern
is a bit more generic, but it should still be pretty fast.)

Note that this change eliminates the need to call sympy.sympify, so
overall this change should lead to much faster code.