Adding Pyomo5 expression walkers
This led to several major clarifications of the code
semantics.  Now the code references walker functions that
with documented semantics.

NOTE:  With this change, the expression walker defined in
pyomo.core.base.template_expr.py is no longer used.  I'll add a
ticket to figure out if we should deprecate that functionality.