Removing the _is_owned flag in expressions
John noted that we probably don't need this flag, which
I confirmed except for the ViewSumExpression class.  That
class needs a flag to indicate whether the argument list
is being shared.  (But that's an internal flag that only
reflects whether two or more ViewSum objects share the
argument list.)

But this change achieves several goals:
. Less memory
. Faster (a little bit)
. No global 'ownership' flag that needs to be set by
  components.

Thus, this change is faster and arguably more robust!