Updates to BARON writer
This is a major change to use a visitor pattern
to write expressions.

But recent changes with expressions and the use
of the symbol map were propigated here.  This led to
significant changes and simplifications in the writer logic.

One particularly noteworthy change is that the writer no longer
walks the tree to find variables.  The variables emitted are
the ones in the model.  BUT, this means that it's OK to generate
a model with variables from (a) another model
and (b) a deactivated block.  This first is an unusual case,
but the latter is not.  Model transformations can easily
create deactivated blocks with variables that we want to continuing
using in our model.  Until we get to a point where variables
are registered "globally", I think we want to allow the use of
variables in this manner.  Further, checking for a "missing" variable
from a model can be quite expensive.