A major change to type checking
I added is_visitor() and is_named_expression() methods,
and I eliminated all global sets of types used to check
expressions, except for a new set:  valid_leaf_types.  This is the
*only* set that developers need to know about to write
a expression visitor.

The key impact of this change is that *all* types in expressions
are well-defined.  The interior nodes in an expression tree
satisfy is_expression(), and all leaf notes have types that
are in valid_leaf_types or which are modeling variables or parameters.