Add a "paraniod" mode to Block.clone()
This fixes #334 by introducing a "paraniod" mode for Block.clone() where
Component.__deepcopy__ makes a copy of the memo before deepcopying the
Component contents.  This avoids the (random) problem where a component
was being partially copied.  This also speeds up clone() for cloning
parts of a model (previously, the whole model was being cloned because
of the _parent weakref).