Make it legal to retrieve a mutable param without a default.
This changes the Param API so that no exception is triggered for
retrieving a _ParamData for a mutable param with no value or default.
This allows users to use mutable params during the model construction
process and not define their values until later.  value(param) still
returns an exception when the `_ParamData._value` is `_NotValid`.