Changing how we verify the GDP chapter examples.
This was a terrible example: it is degenerate, and there is no objective
pressure on the optimization variables.  We finally got "bit" by the
reference solver returning a different - but equally valid - answer.
This moves the test from a simple diff to a more general check that the
solution returned was a valid solution to the example.