Changing how quadratic ROM geometries are cached.
Moving to an autogenerated / updated .py file to cache quadratic ROM
geometries instead of individual ".out" files.  This is more compatible
with python packaging and removes dependence on the filesystem paths,
except when updating the cache.  Also improving the generator so that it
will update the cache instead of starting from scratch.