Reworking design for 'default ordering'.
The previous edits confused set ordering with set storage.

This commit clarifies that the default behavior in Set() is to
have a set that has no declared ordering, but for which the
iteration order is fixed.  The data() method still returns a set(),
and insertion checks are still fast.  But discard() might be slower,
though.