Removing the dict API from slices.
It turns out that it is not obvious what the dict API should be.  In
some cases, you want the keys to be in the context of the original
indices and in other cases you need it in the context of only the
wildcards.  This change forces users to be explicit about what they
want.