Add new example
If no external function is defined, runtime exception is thrown.
This should really be converted to a test in the TestPyomoInterface.py.
I am not even sure this is the proper way to handle this problem.
Previously, it was erroring with an error when converting from type
array[0] to type shape(0).