Adding method to return F,G,H from AMPL external functions
This adds a public method to AMPLExternalFunction that allows a user to
obtain the first and second derivative information from the external
function.  This now only passes non-NULL derivs and hes arrays to the
external function if the derivatives are actually requested, and adds
the ability to annotate some variables as "fixed" (necessary for some
GSL test functions).  This fixes #664.