Adding new factory for executables
Currently, Pyomo uses the pyutilib.services logic for
executables.  This searches immediately when executables are registered.
While this may be desired in some contexts, within Pyomo this is
undesirable since executables are searched for when Pyomo is imported
even if those executables are never used.

This new factory delays the search for an executable until its
path is requested.