Added tests for ConfigBlocks
ConfigBlocks are now working as expected. However there is still a bug in the code that
throws an exception when calling the solve function a subsequent time. I separated
the tests into six files for now to show that individually the ConfigBlock code is
doing what is expected. Once I fix the bug I will remove the individual files and only
keep TestTRConfig.py.