Adding logic to control signals in shellcmd solvers
By default, Pyomo calls PyUtilib to launch solvers using signals.
This causes problems in some multithreaded applications, where Pyomo
is not running in the main thread.  We added control of signal
management to PyUtilib last year, and this branch propagates this
change to Pyomo, using the flag 'use_signal_handling' (which defaults
to True).