Remove Port.Conservative, be more selective for evar bounds
This removes the Conservative rule and its helper methods and instead
changes the handling of include_splitfrac to that specifying
include_splitfrac=False will prevent the split fraction variables from
being created.  This also changes how expanded arc variables' domains
and bounds are set, so that the domain is only tightened when the arc
has a single source or single destination.