:zap: `get_values()` efficiently
Asking the `cplex` solution for *specific* variables' values is extremely slow due to the conversion on `cplex`'s end to lookup the variables you've asked for.
It is orders of magnitude faster to get the full solution vector. Even worse, using the "specific variable interface" to get the full solution vector.
If we must get specific variables (ie. `_load_vars()`) then their index should be used instead of their name.