fix-ph-convert-string-to-number
The comment `ph.py` suggests that an attempt is made to convert string arguments to numbers. It says that integer is tried first, and then float. However, the implementation is opposite and consequently all integer arguments are converted to floats. This has caused a problem in my CBC model when I try to pass an integer value.

The changes here switch the order of the conversion attempts to reflect the written comment and allow integer arguments to be passed correctly.