Rework Pyomo logger to use Formatters and defer to global loggers
The reworks the Pyomo logger to more closely adhere to the python
logging conventions while maintaining backwards compatibility.  It moves
much of the logic out of the LogHandler class and into the
WrappingFormatter.  It better bahaves with other systems by suppressing
pyomo log output if a root log handler is defined.