Improving 'pyomo solve' logic
There has been a long-standing bug in the 'pyomo solve' command.
Specifically, it fails to execute properly when using a solver manager
that is executed remotely (e.g. the 'neos' solver manager) when the
solver is not installed locally.

This change is a bit of a hack to fix that.  The 'asl' solver is used
to process the command-line arguments in that case, which are sent to the
solver manager.