Overhaul @deprecated to work better with class, enum, and sphinx
This is an overhaul of the @deprecated decorator to improve how the
decorator works with classes and enum types.  In particular, decorating
a class will automatically decorate the __init__() method.  This way
both the class docs and the user warning will be set.  This also updates
the docstring to be in sphinx format.