Improve comparison of Book examples to baselines
This reworks test_book_examples:compare_files to ignore extra entries in
the output that look like values that are very clode to zero.  This
fixes an issue where different versions of solvers occasionally return
slightly nonzero values for variables that are nominally zero (within
tolerance) in the reference solution.