extra root + tree processing times and add to solver results
Had to convert TextIOWrapper directly to a string so that we can also check multi-line statements.
Also, for CPLEX versions before 12.5.1 it seems like log parsing was not supported; this should
not matter for us (but this is made more explicit here by trying to parse an empty string instead of
the file name itself). This could be cleaned up further in the future, but I am trying to
keep the diff small for now.