Improved results object when SCIP is used.
This commit seeks to correct issues when using the SCIP solver, namely concerning the results object.

The results object now includes information about the optimisation process, as found in the log file. This includes the time needed for the optimisation (in seconds; as computed by SCIP itself, not Pyomo), the (MIP) gap (in %), the primal bound and the dual bound. This is now found under the 'Solver' key.

The SolverStatus and TerminationCondition attribute combinations have also been revised to avoid warnings. 

Other problems remain. The 'Problem' part of the results object is updated in ways that were not understood. As such, no changes were made to that dict, but it is logical to use the 'Solver' dict to update the lower and upper bounds. Overall, the interaction with solvers does not appear to be sufficiently standardised, let alone documented, making pyomo (way) less appealing than it should be. It does work, of course, but making changes to it is not for the faint of heart.

One final note about getting information from SCIP: it can be asked to generate  a statistics file, which contains all the information one is likely to need to fill the results object with data. Since that would involve messing with the SCIP command, I decided not to do it for the time being.