Fix: mosek_direct updated to use putqconk instead of putqcon
This fix concerns QCQP models when solved using mosek. In MOSEK's
Optimizer API, the putqcon method resets the Q matrix entries for
all constraints to zero, while putqconk does so only for k-th
constraint. The _add_constraints method in mosek_direct would
call putqcon, but this would lead to loss of Q info with every
subsequent call to the _add_constraints (if new Q info was given).
This is now fixed, because the Q matrix in each constraint is
updated in its own call to putqconk.