Set LogToConsole instead of OutputFlag to handle "tee"
In Gurobi, there are three parameters related to logging:

LogToConsole is a boolean flag that toggles whether the solver output is logged
to the console. (This fits the intention behind setting `tee` in Pyomo.)

LogFile is a string that refers to a file name. If non-empty, solver output will
be appended to that file.

Finally, OutputFlag is a combination of the two: If set to 0, both LogToConsole
is set to 0, as well as LogFile set to an empty string.

In the old implementation, if a user has already specified a LogFile via the
`options`, but then calls `solve(tee=False)`, the log files will not be written.
This change makes sure that `tee` only affects what happens to the console.

Only GurobiDirect and the derived GurobiPersistent have this logic implemented
that way. For the `appsi_gurobi` implementation, `tee` already had no effect on
what happened with log files.